requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
yt-dlp
openai-whisper
torch
transformers
numpy
opencv-python
langchain-deepseek
langchain-core
loguru
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu121 